requirements.txt
attrs==22.1.0
black==22.6.0
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
colored==1.4.3
emoji==2.0.0
idna==3.4
iniconfig==1.1.1
mypy-extensions==0.4.3
nanoid==2.0.0
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.2
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pytest==7.1.2
PyYAML==6.0
requests==2.28.1
simple-term-menu==1.5.0
tomli==2.0.1
typing_extensions==4.3.0
urllib3==1.26.12
